Refactor button components for consistency and accessibility; update Cart and FeedbackModal for improved UX. Add logging to section management in sectionsStore and PracticePlanForm. Adjust error handling in edit action for better clarity.

+++ src/lib/server/services/practicePlanService.js
@@ -796,7 +796,7 @@ export class PracticePlanService extends BaseEntityService {
 
 				const updateQuery = `
           UPDATE practice_plan_drills 
-          SET order_in_plan = ${caseStatement} 
+          SET order_in_plan = (${caseStatement})::integer
           WHERE section_id = $1 AND id = ANY($${values.length + 1}::int[])`;
 
 				// Add the array of item IDs as the last parameter
